email-rewriting-tool-extension-/requirements.txt
fastapi
uvicorn
google-generativeai
pydantic
